% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMR_func.R
\name{RBMR_func}
\alias{RBMR_func}
\title{RBMR function for only one LD block}
\usage{
RBMR_func(
  bh1,
  bh2,
  se1,
  se2,
  gamma,
  alpha,
  sgga2,
  sgal2,
  beta0,
  R,
  alphag,
  betag,
  constr,
  epsStopLogLik,
  maxIter
)
}
\arguments{
\item{bh1}{A vector of SNP effects on the exposure variable, usually obtained from a GWAS.}

\item{bh2}{A vector of SNP effects on the outcome varible, usually obtained from a GWAS.}

\item{se1}{A vector of standard errors of \code{bh1}.}

\item{se2}{A vector of standard errors of \code{bh2}.}

\item{gamma}{A vector to intialize the true value of the mean of the exposure.}

\item{alpha}{A vector to intialize the true value of the mean of the direct effect.}

\item{sgga2}{A constant to initialize the true value of the standard error of exposure \code{bh1}.}

\item{sgal2}{A constant to initialize the true value of the standard error of direct effect \code{bh2}.}

\item{beta0}{A constant, initialize the true value of causal effect.}

\item{R}{A matrix, LD matrix.}

\item{alphag}{A constant to initialize the parameter of multivariate generalize t distribution}

\item{betag}{A constant to initialize the parameter of multivariate generalize t distribution}

\item{constr}{0 or 1, when constr is equal to 0, the function calculates the ELBO under the alternative hypothesis, when constr is equal to 1, the function calculates the ELBO under the null hypothesis.}

\item{epsStopLogLik}{Numerical precision.}

\item{maxIter}{Maximum number of interations to solve the estimating equations.}
}
\value{
A list
\describe{
\item{beta0}{Estimated causal effect.}
\item{sgal2}{Standard error of the direct effec.}
\item{sgga2}{Standard error of the exposure.}
\item{Iteration}{Number of iterations.}
\item{loglik}{The result of loglikelihood at convergence.}
\item{diff}{The difference between the value of loglikelihood at convergence and the value of the loglikelihood of previous itertaion.}
\item{tstat}{The value of the ELBO.}
\item{alpha_w}{The parameter of multivariate generalized t distribution.}
\item{beta_w}{The parameter of multivariate generalized t distribution.}
}
}
\description{
When the number of LD block is 1, this function estimates the causal effect.
}
