% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBMR_func.R
\name{RBMR_func}
\alias{RBMR_func}
\title{main function for RBMR (only one block)}
\usage{
RBMR_func(
  bh1,
  bh2,
  se1,
  se2,
  gamma,
  alpha,
  sgga2,
  sgal2,
  beta0,
  R,
  alphag,
  betag,
  constr,
  epsStopLogLik,
  maxIter
)
}
\arguments{
\item{bh1}{A vector of SNP effects on the exposure variable, usually obtained from a GWAS}

\item{bh2}{A vector of SNP effects on the outcome varible, usually obtained from a GWAS}

\item{se1}{A vector of standard errors of \code{bh1}}

\item{se2}{A vector of standard errors of \code{bh2}}

\item{gamma}{A vector, intialize the true value of the mean of the exposure}

\item{alpha}{A vector, intialize the true value of the mean of the direct effect}

\item{sgga2}{A constant, initialize the true standard error of exposure \code{bh1}}

\item{sgal2}{A constant, initialize the true standard error of direct effect \code{bh1}}

\item{beta0}{A constant, initialize the true value of causal effect}

\item{R}{A matrix, LD matrix}

\item{alphag}{A constant, initialize the parameter of multivariate generalized t-distribution}

\item{betag}{A constant, initialize the parameter of multivariate generalized t-distribution}

\item{constr}{0 or 1, 0 represents the likelihood ratio test of alternative hypothesis, 1 represents the likelihood ratio test of null hypothesis.}

\item{epsStopLogLik}{Numerical precision}

\item{maxIter}{The maximum number of iterations}
}
\value{
A list
\describe{
\item{beta0}{Estimated causal effect}
\item{sgal2}{Standard error of direct effect}
\item{sgga2}{Standard error of exposure}
\item{Iteration}{Iteration}
\item{loglik}{the result of loglikelihood at convergence}
\item{diff}{The difference between the value of loglikelihood at convergence and the value of the loglikelihood of previous itertaion}
\item{tstat}{The value of the loglikelihood ratio test}
\item{alpha_w}{The parameter of multivariate generalized t distribution}
\item{beta_w}{The parameter of multivariate generalized t distribution}
}
}
\description{
main function for RBMR (only one block)
}
